AI-Tripper-backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx

# Database
sqlalchemy
asyncpg
greenlet  # SQLAlchemy async için gerekli
alembic

# Redis (optional)
redis
hiredis

# Authentication
bcrypt==4.2.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Validation
email-validator
pydantic[email]